Move `ImmuTable` plus a few dependencies to the `ImmuTable` package
