+++ Modules/Sources/ImmuTable/CellRegistrar.swift
@@ -1,0 +1,3 @@
+protocol CellRegistrar {
+    func register(_ cell: ImmuTableCell, cellReuseIdentifier: String)
+}
