+++ WordPress/Classes/ViewRelated/Activity/ActivityListViewModel.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import WordPressFlux
 
 protocol ActivityPresenter: AnyObject {
