+++ WordPress/Classes/ViewRelated/Me/App Settings/MediaCacheSettingsViewController.swift
@@ -1,4 +1,5 @@
 import Foundation
+import ImmuTable
 
 class MediaCacheSettingsViewController: UITableViewController {
     fileprivate var handler: ImmuTableViewHandler?
