+++ WordPress/Classes/ViewRelated/Plugins/Controllers/PluginListViewController.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 import WordPressKit
 import WordPressFlux
